Printable PDF

+++ app/Http/Controllers/DeckPdfController.php
@@ -14,7 +14,7 @@ class DeckPdfController extends Controller
     public function download(Request $request, Deck $deck)
     {
         // Ensure the deck belongs to the authenticated user
-        abort_unless($deck->creator_id === auth()->id(), 403);
+        abort_unless((int) $deck->creator_id === (int) auth()->id(), 403);
 
         return app(DeckPdfService::class)->download($deck);
     }
